+++ addons/dice_roller/dice_def.gd
@@ -3,14 +3,26 @@ class_name DiceDef
 extends Resource
 
 @export var name: String
-@export var color: Color = Color.ANTIQUE_WHITE
-@export_enum("D6:6", "D4:4", "D8:8", "D10:10", "D10x10:100", "D20:20") var sides = 6
 
-const icons =  {
-	4: preload("res://addons/dice_roller/dice/d4_dice/d4_dice.svg"),
-	6: preload("res://addons/dice_roller/dice/d6_dice/d6_dice.svg"),
-	8: preload("res://addons/dice_roller/dice/d8_dice/d8_dice.svg"),
-	10: preload("res://addons/dice_roller/dice/d10_dice/d10_dice.svg"),
-	100: preload("res://addons/dice_roller/dice/d10x10_dice/d10x10_dice.svg"),
-	20: preload("res://addons/dice_roller/dice/d20_dice/d20_dice.svg"),
-}
+@export var color: Color = Color.ANTIQUE_WHITE
+
+@export var shape: DiceShape: #  = DiceShape.new("D6")
+	set(value):
+		print ("DiceDef setting shape to: ", value)
+		shape=value
+	get():
+		return shape
+## @deprecated use `shape` instead
+
+@export_storage var sides: int:
+	set(value):
+		print("DiceDef setting sides to ", sides, " for ", name)
+		if value:
+			shape = DiceShape.from_sides(value)
+	get():
+		# zero means migrated
+		print("asking for sides")
+		return 0
+
+func _to_string() -> String:
+	return "DiceDef('"+ name + "' " + str(shape) + " "  + str(color) + ")"