+++ examples/dice_roler/dice_set_editor.gd
@@ -2,13 +2,14 @@ class_name DiceSetEditor
 extends ConfirmationDialog
 
 @onready var tree : Tree = $Tree
-@onready var nsides_popup := $SidesPopup
+@onready var shapes_popup := $SidesPopup
 @onready var color_popup := $ColorPopup
 @onready var color_picker := $ColorPopup/ColorPicker
 @onready var preset_name_edit := $PresetNameDialog/MarginContainer/VBoxContainer/PresetNameEdit
 
+const delete_icon := preload("./delete-icon.svg")
+const color_icon := preload('./color-icon.svg')
 const diceset_dir = "user://dicesets_v1/"
-
 var load_button : Button
 var save_button : Button
 var preset_list : Array[String] = []
@@ -20,7 +21,7 @@ enum {
 
 enum {
 	COL_ICON,
-	COL_SIDES,
+	COL_SHAPE,
 	COL_COLOR,
 	COL_ID,
 	COL_REMOVE,
@@ -85,6 +86,7 @@ func load_preset(preset_id: int):
 	var config = ConfigFile.new()
 	config.load(diceset_dir.path_join(preset_name + ".diceset"))
 	var dice_set = config.get_value('default','dice_set')
+	print_rich(dice_set)
 	if not dice_set: return
 	set_dice_set(dice_set)
 
@@ -92,31 +94,30 @@ func load_preset(preset_id: int):
 ## Dice list methods
 
 func setup_dice_list():
-	for nsides in DiceDef.icons:
-		nsides_popup.add_icon_item(DiceDef.icons[nsides], str(nsides), nsides)
-	nsides_popup.id_pressed.connect(_on_popup_nsides_id_pressed)
-	
+	for shape in DiceShape.options():
+		shapes_popup.add_icon_item(DiceShape.icon_for_shape(shape), shape)
+	shapes_popup.index_pressed.connect(_on_shape_popup_index_pressed)
 	color_picker.color_changed.connect(_on_color_picker_changed)
 
 	var _root := tree.create_item()
 	tree.set_column_expand(COL_ICON, 0)
-	tree.set_column_title(COL_SIDES, "Sides")
-	tree.set_column_expand(COL_SIDES, 0)
+	tree.set_column_title(COL_SHAPE, "Sides")
+	tree.set_column_expand(COL_SHAPE, 0)
 	tree.set_column_title(COL_COLOR, "Color")
 	tree.set_column_expand(COL_COLOR, 0)
 	tree.set_column_title(COL_ID, "Name")
 	tree.button_clicked.connect(_on_tree_button_clicked)
 	tree.item_edited.connect(_on_tree_item_edited)
 	add_add_row()
-	add_dice_row(6, Color.CORAL, 'coral dice')
-	add_dice_row(4, Color.FIREBRICK, 'red dice')
+	add_dice_row(DiceShape.new("D6"), Color.CORAL, 'coral dice')
+	add_dice_row(DiceShape.new("D4"), Color.FIREBRICK, 'red dice')
 
 func _on_tree_button_clicked(item: TreeItem, _column: int, id: int, _mouse_button_index: int):
 	match id:
 		BUTTON_DELETE:
 			item.free()
 		BUTTON_ADD:
-			add_dice_row(6, Color.ANTIQUE_WHITE, "Dice")
+			add_dice_row(DiceShape.new("D6"), Color.ANTIQUE_WHITE, "Dice")
 
 func _on_tree_item_edited():
 	var item := tree.get_edited()
@@ -127,9 +128,9 @@ func _on_tree_item_edited():
 			color_picker.color = item.get_metadata(COL_COLOR)
 			var target_rect = tree.get_screen_transform()*tree.get_custom_popup_rect()
 			color_popup.popup_on_parent(target_rect)
-		COL_SIDES:
+		COL_SHAPE:
 			var target_rect = tree.get_screen_transform()*tree.get_custom_popup_rect()
-			nsides_popup.popup_on_parent(target_rect)
+			shapes_popup.popup_on_parent(target_rect)
 
 func _on_color_picker_changed(color: Color):
 	var item := tree.get_edited()
@@ -137,12 +138,11 @@ func _on_color_picker_changed(color: Color):
 	item.set_icon_modulate(COL_COLOR, color)
 	item.set_icon_modulate(COL_ICON, color)
 
-func _on_popup_nsides_id_pressed(id):
-	var nsides = str(id)
+func _on_shape_popup_index_pressed(index):
+	var shape_name = shapes_popup.get_item_text(index)
 	var item := tree.get_edited()
-	item.set_metadata(COL_SIDES, id)
-	item.set_text(COL_SIDES, nsides)
-	item.set_icon(COL_ICON, DiceDef.icons[id])
+	item.set_text(COL_SHAPE, shape_name)
+	item.set_icon(COL_ICON, DiceShape.icon_for_shape(shape_name))
 
 func add_add_row():
 	var root := tree.get_root()
@@ -150,22 +150,23 @@ func add_add_row():
 	item.set_expand_right(COL_ICON, true)
 	item.add_button(COL_ICON, preload("./add-icon.svg"), BUTTON_ADD)
 
-func add_dice_row(nsides: int, color: Color, id: String):
+func add_dice_row(shape: DiceShape, color: Color, id: String):
+	var nsides := DiceShape.to_sides(shape)
+	print("Adding dice row: d", nsides, " ", shape, " ", color, " ", id)
 	var root := tree.get_root()
 	var item := tree.create_item(root, root.get_child_count()-1) # left the add row the last
-
-	item.set_icon(COL_ICON, DiceDef.icons[nsides])
+	
+	item.set_icon(COL_ICON, shape.icon() if shape else null)
 	item.set_icon_max_width(COL_ICON, 32)
 	item.set_icon_modulate(COL_ICON, color)
 
-	item.set_cell_mode(COL_SIDES, TreeItem.CELL_MODE_CUSTOM)
-	item.set_metadata(COL_SIDES, nsides)
-	item.set_text(COL_SIDES, str(nsides))
-	item.set_editable(COL_SIDES, true)
-	item.set_text_alignment(COL_SIDES, HORIZONTAL_ALIGNMENT_CENTER)
+	item.set_cell_mode(COL_SHAPE, TreeItem.CELL_MODE_CUSTOM)
+	item.set_text(COL_SHAPE, shape.to_string() if shape else "--")
+	item.set_editable(COL_SHAPE, true)
+	item.set_text_alignment(COL_SHAPE, HORIZONTAL_ALIGNMENT_CENTER)
 
 	item.set_cell_mode(COL_COLOR, TreeItem.CELL_MODE_CUSTOM)
-	item.set_icon(COL_COLOR, preload('./color-icon.svg'))
+	item.set_icon(COL_COLOR, color_icon)
 	item.set_icon_modulate(COL_COLOR, color)
 	#item.set_custom_bg_color(COL_COLOR, color)
 	item.set_metadata(COL_COLOR, color)
@@ -174,7 +175,7 @@ func add_dice_row(nsides: int, color: Color, id: String):
 	item.set_text(COL_ID, id)
 	item.set_editable(COL_ID, true)
 
-	item.add_button(COL_ID, preload("./delete-icon.svg"), BUTTON_DELETE, false, "Remove")
+	item.add_button(COL_ID, delete_icon, BUTTON_DELETE, false, "Remove")
 
 func get_dice_set() -> Array[DiceDef]:
 	var dice_set: Array[DiceDef] = []
@@ -185,7 +186,7 @@ func get_dice_set() -> Array[DiceDef]:
 		var def = DiceDef.new()
 		def.name = item.get_text(COL_ID)
 		def.color = item.get_metadata(COL_COLOR)
-		def.sides = item.get_metadata(COL_SIDES)
+		def.shape = DiceShape.new(item.get_text(COL_SHAPE))
 		dice_set.append(def)
 	#print(dice_set)
 	return dice_set
@@ -200,4 +201,4 @@ func clear_dice_set():
 func set_dice_set(dice_set : Array[DiceDef]):
 	clear_dice_set()
 	for dice in dice_set:
-		add_dice_row(dice.sides, dice.color, dice.name)
+		add_dice_row(dice.shape, dice.color, dice.name)